Refactored to pull plugin logic  together
