+++ extensions/azurePublishNew/src/node/azureResources/constants.ts
@@ -1,0 +1,12 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+export const AZURE_HOSTING_GROUP_NAME = 'App Services';
+export const COGNITIVE_SERVICES_GROUP_NAME = 'Cognitive Services';
+
+export const SO_STANDARD_TIER = 'S0 Standard';
+export const S1_STANDARD_TIER = 'S1 Standard';
+export const STANDARD_LRS_TIER = 'Standard LRS';
+export const PAY_AS_YOU_GO_TIER = 'Pay as you go';
+export const FREE_APP_REGISTRATION_TIER = 'Free'; // aka included, N/A
+export const FREE_COGNITIVE_SERVICES_TIER = 'F0';
