+++ extensions/azurePublishNew/src/node/index.ts
@@ -1,51 +1,91 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import { IExtensionRegistration } from '@botframework-composer/types';
+import { IBotProject, IExtensionRegistration, ProcessStatus, PublishPlugin } from '@botframework-composer/types';
 
-// TODO: remove this code eventually, added it so that extension would successfully build
+import { availableResources, setUpProvisionService } from './availableResources';
+import { BackgroundProcessManager } from './backgroundProcessManager';
+import { GetResourcesResult, ProvisionConfig, ResourceDefinition, ResourceProvisionService } from './types';
 
-function initialize(registration: IExtensionRegistration) {
-  const plugin1 = {
+/**
+ * Creates the azure publishing plug-in for this extension.
+ * @returns The plug-in with properties and methods.
+ */
+const createAzurePublishPlugin = (): PublishPlugin<ProvisionConfig> => {
+  const getResources = (project: IBotProject /*user?: UserIdentity*/): Promise<GetResourcesResult[]> => {
+    const provisionServices: Record<string, ResourceProvisionService> = {};
+
+    return Promise.resolve(
+      availableResources.reduce((resources: GetResourcesResult[], currentResource: ResourceDefinition) => {
+        const service = provisionServices[currentResource.key];
+        if (service.getRecommendationForProject(project) !== 'notAllowed') {
+          resources.push({
+            ...currentResource,
+            required: service.getRecommendationForProject(project) === 'required',
+          });
+        }
+        return resources;
+      }, [])
+    );
+  };
+
+  const getStatus = async (/* config, project, user */) => {
+    const response = {
+      status: 200,
+      result: {
+        time: new Date().toString(),
+        message: 'Publish successful.',
+        log: '',
+      },
+    };
+    return response;
+  };
+
+  const provision = async (
+    config: ProvisionConfig,
+    project: IBotProject
+    // user?: UserIdentity,
+    // getAccessToken?: GetAccessToken
+  ): Promise<ProcessStatus> => {
+    const jobId = BackgroundProcessManager.startProcess(202, project.id, config.key, 'Creating Azure resources...');
+    setUpProvisionService(config);
+    return BackgroundProcessManager.getStatus(jobId);
+  };
+
+  const publish = async (_config, _project, metadata /*_user */) => {
+    const response = {
+      status: 202,
+      result: {
+        time: new Date().toString(),
+        message: 'Publish accepted.',
+        log: '',
+        comment: metadata.comment,
+      },
+    };
+    return response;
+  };
+
+  return {
+    bundleId: 'publish', // we have custom UI to host
     name: 'azurePublishNew',
     description: 'Publish bot to Azure - new',
-    bundleId: 'publish' /** we have custom UI to host */,
-    publish,
+    getResources,
     getStatus,
+    provision,
+    publish,
   };
-  registration.addPublishMethod(plugin1);
+};
+
+//----- Register the plug in -----//
+const initialize = (registration: IExtensionRegistration) => {
+  registration.addPublishMethod(createAzurePublishPlugin());
 
   // test reading and writing data
   registration.log('Reading from store:\n%O', registration.store.readAll());
 
   registration.store.replace({ some: 'data' });
   registration.log('Reading from store:\n%O', registration.store.readAll());
-}
-
-async function getStatus(config, project, user) {
-  const response = {
-    status: 200,
-    result: {
-      time: new Date().toString(),
-      message: 'Publish successful.',
-      log: '',
-    },
-  };
-  return response;
-}
-
-async function publish(config, project, metadata, user) {
-  const response = {
-    status: 202,
-    result: {
-      time: new Date().toString(),
-      message: 'Publish accepted.',
-      log: '',
-      comment: metadata.comment,
-    },
-  };
-  return response;
-}
+};
 
 module.exports = {
   initialize,
